backend/pyproject.toml
[tool.poetry]
description = ""
authors = ["Saliovin <jonas.eukan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
email-validator = "^2.2.0"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
httpx = "^0.27.0"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
